backend/core/requirements.txt
amqp==5.0.6
anyio==3.3.2
asgiref==3.4.1
billiard==3.6.4.0
blinker==1.4
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.4
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
et-xmlfile==1.1.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==2.10
kombu==5.1.0
MarkupPy==1.14
nplusone==1.0.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rfc3986==1.5.0
sentry-sdk==1.3.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
tablib==3.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
pytest-django==4.5.2
django-admin-rangefilter==0.8.3
django-modern-rpc==0.12.1
sentry-sdk
pycryptodome==3.14.1
gcloud==0.18.3
django-summernote==0.8.20.0
